use -XTypeApplications for 'focus' in keybindings
Wanting to relax the return type of the 'focus' action from
'AppState' to '()' revealed an opportunity to improve usability.  In
keybindings, when 'focus' is used we currently use an explicit,
fully saturated type signature to determine the Focusable instance.
This includes the 'Action' constructor and the 'AppState' return
type, which are both redundant (`chain` and `chain'` both ignore the
Action return type as a result of the StateT refactor).

Therefore we can simplify the definition of keybindings that use
'focus' by using -XTypeApplications, available in GHC since 8.0.