nix: Adopt to renamed test option in cabal-install 3.0
The nix release we've chosen already ships the cabal-install tool
version 3.0.

This version has renamed a test command we use to stream in-progress
test output to the console. This patch adopts the new test command and
furthermore uses the `v2` test API. Reason being is that locally, there
seemed to be no difference between test and new-test, while v2-test
seemed to be the new command.