Improve help for setTags action
The 'apply given tags' isn't very helpful, particularly when we have a
look at our help page in the UI:

     a - apply given tags
     S - apply given tags

Both are distinct actions and the 'what' is missing here. This adds at
least the show instances of the tag operations to make this part of the
UI more helpful.