curry sendmail effect with executable
We have a sendmail executable path field in the config.  But this
isn't really needed, and indeed makes no sense for alternative
implementations (such as the "write to file" implementation we use
for the test suite, evident in that the argument is ignored).

So remove the field, and partially apply the sendmail function to
the executable path in the cvSendMailPath field.