avoid construction of strict ByteStrings
When sending or saving mail, we currently render the mail into a
strict ByteString.  This could be quite an expensive operation when
all we really need to do is send it to sendmail(1) stdin or to a
file.  Change the cvSendMailCommand to accept a Builder instead of
a strict ByteString, allowing a more efficient implementation, and
chase the changes.

Also remove the indexMail function.  It should not be the
responsibility of the Notmuch subroutines to actually write the mail
to disk (and pushing that responsibility to the caller allows the
caller to use the most efficient means of writing the file).