Remove Nix ghcHEAD build
I honestly don't know what motivation I head adding it in the first
place. The build has been failing for months and there is currently no
motivation to look into any of those failures. At this point part of the
failing ghcHEAD Nix job are dependencies which can not be build. All in
all no point of keeping it around.