Raise an error if setTags is called on unsupported widget
If we call `setTags` on a widget we have not registered a handler for,
Purebred would simply assume that the only widget would be the list of
threads.

However when adding new widgets, this can lead to confusing
behaviour for new contributors since the handler is silently operating
on the list of threads.