Operate on toggled *or* selected list items
This adds an additional helper for lists supporting toggling of it's
items. It is similar in nature to our existing `selectedItemHelper`.

It really is a workaround for Brick's List, since it can not encapsulate
a marked/toggled state for it's list items. This function allows to
operate on toggled or selected items.