Use a better word than "Tagged"
I had used "Tag", because mutt is using "Tag". The word tagging here has
a different context however. In mutt it used to be to perform bulk
actions on the tagged list items, but with the introduction of notmuch,
tagging can also mean the labeling of mails with "tags".

Instead of using "Tagged", use the word "Marked" since we're marking
these list items for further action.