Make the colour of a list item consistent
Through different list item states, the colours of each
part (e.g. authors, tags, etc) were not consistent. The authors for
example, would show up with a white background on gray text, while the
entire listitem would show the inverse colours.

The reason for this is that we only carry information about the state of
the list item into the list item renderer. We selected different
attributes depending on each state. These attributes were not different
for other parts of the list item, e.g. the tags or the authors.

With the introduction of an additional state (toggling), it was become
more cumbersome to keep track of the states and selecting their
attributes. With the combinatorial explosion of possibilities we had to
find a different way.

This patch replaces the old code of selecting attributes with a way of
generating all parts of the list with the same calculated tree of
attributes.

We use a base attribute and concatenate the additional states on to the
base attribute depending whether the state is true or not. For example
if we're dealing with a new thread which is toggled, the attribute would
be:

        list.toggled.newItem

If it's selected, but not toggled, but new:

        list.selected.newItem

or a selected and toggled item:

        list.selected.toggled

and so forth.

The order the states are concatenated is honoured. Which means all
colour definitions for the base item, e.g.:

       list

are the same for:

        list.selected.toggled.newItem

if not overwritten.

Same goes for tags and mail attributes. This way everyone can define
colours to their hearts content for each state the items could be in (27
possible combinations if I'm not wrong) or keep the default.