Use an explicit way to update the attachments
The way we updated the attachment list was by guessing the
operation. Well, if the list was empty the operation was pretty straight
forward: insert the item(s). However if the list was not empty the
guessing became harder: either replace or append. The ambiguity of this
resulted in this bug.

Instead of guessing, pass the operation we want to do after we've
invoked the operator to the function which invokes the editor. We know
exactly what we want to do with the list, after the editor has exited.

Fixes https://github.com/purebred-mua/purebred/issues/379