Use a tool to dump keybindings
This tool can be used to dump our key bindings as Asciidoc markup. The
markup can be used to be incorporated in our documentation.

Reason as to why this would be useful: part of customising our
keybindings is knowing the existing keybindings and what key event they
are registered for. This way, there is no need to dig through the code
to figure out what is registered for what event.