Make the search path editor consistent with other editors
This adds additional key bindings to make the editor consistent with
other input widgets in Purebred.

Before using the search path editor was very confusing, since the only
way to move the focus back to the list (a.k.a file browser) was by
pressing Escape.

This patch changes the behaviour to allow the focus change to the list
with more key strokes.