Make the threads view constructor more concise
This views former name was `MailView` because it was keeping state of
the list of mails and threads.

This is no longer true. The view which shows the threads mails is a
separate type. In order to avoid any confusion, rename the type to what
state we're actually holding here and that is the list of threads.