travis: Stream test progress
This streams the test output to the console in CI. This may be a
solution for #321 if the underlying cause is that no test output is
directed to stdout until the test command finishes.

If however the cause is something different, this patch may at least
shed some light into why CI kill's the job without output for 10min.

This patch comes with a small caveat tho. We default to the `new-test`
command. However, streaming support existed for older cabal versions
provided by the test command. Newer cabal versions (>= 3) provide the
streaming support with the `v2-test --test-show-details` command. This
is not supported with previous cabal-install releases. The old `test`
command in return does not understand the `--enable-tests` and
`--enable-benchmarks` options.

TLDR; there are too many incompatibilities between cabal-install
versions and their options, that streaming is only enabled for GHC >
8.6.5 builds with newer cabal-install versions.

Relates to https://github.com/purebred-mua/purebred/issues/321