Use new widget to keep the editor state
This adds a new editor widget in the Brick namespace. The widget is a
simple wrapper around the edit widget from Brick. However it can save
it's state and revert back to it. This is especially useful, if we want
to cancel input and not commit it to Purebred.

Note though, that some editors have not been changed to StatefulEditor
widgets. Reason being is that those don't really need to keep state. For
example the editors which change the tags on threads or mails. User's
supply operations. If they're empty, no tag operations are
performed. The editor contents are reset with every focus change.

Fixes https://github.com/purebred-mua/purebred/issues/396