Handle garbage path input
This patch fixes a problem with the path editor setting the working
directory in the file browser. Since the file browser itself is not
checking whether the working directory is valid, it is up to the caller
to verify we're not setting anything invalid.

This patch checks whether the path for the working directory exists and
otherwise shows an error.

Fixes https://github.com/purebred-mua/purebred/issues/391