Remove test for file browser
This must have been a mistake. The file browser currently does not use
any type of validation, but uses the same stateful editor. However the
assert function asserts that the error set by a validation function is
cleared, which may never happen. The clearing of errors currently only
happens on keypress either which isn't happening in the assertion as
well. In summary, until we can use validation functions performing IO
actions, don't use this for now.