Workaround for composeAsNew
Initially composeAsNew was implemented with the idea to continue draft
mails. The mails are stored in the file system and indexed in
notmuch. Once the draft is picked up again, the mail is unindexed and
removed from the file system until it's sent.

However, if we use the same function on existing mails in order to
use the mail as a template for a new mail, we actually unindex and
remove the mail mistakenly from notmuch and the file system. Instead all
we want is a copy for composing.

This is a workaround for now, since ideally we want to keep all
compositions as a draft on the file system (see #351).

This patch only unindexes/removes draft mails, any other mail contents
are simply copied for composition.

Fixes https://github.com/purebred-mua/purebred/issues/395