Fall back to previous view in case of editor error
When Purebred is suspended and we hand control over to our editor, it is
possible that the editor errors (crashes, doesn't start, wrong editor,
yadda yadda).

Currently Purebred would simply go to the compose editor with an empty
attachment. That means that we could possibly send an email with an
empty body.

Instead, parameterize the `invokeEditor` function in which we give a
fallback view and widget. In cases of an error, we fall back to this
view and the focused widget.

Fixes: https://github.com/purebred-mua/purebred/issues/336