nix: use new release and GHC 8.8.4
This pin's to a new nix release to avoid CI mostly fetching cached
packages instead of rebuilding the (Nix) world.

With bumping minimum versions for vty and brick, the minimum boundaries
of all transitive packages has shifted upwards, causing a lot of
derivations to be rebuilt due to the fact that they are not included in
the package closure of the former pinned versions.