Keep the tempfile when opening attachment
When opening the attachment with an external command, Purebred should
keep the temporary file it creates. Not all external commands exit after
the file has been seen by the user. Firefox for example, uses a wrapper
which hands the HTML file to the running instance and immediately
exits. That causes the running Firefox instance to open a deleted file.