Make tests using "Not" more robust
This uses what we've learned in the previous change and applies it to
the remaining tests using a "Not" condition without asserting the
presence of the element in the first place. This only applies to
asserting that something was present in the previous screen and is now
gone in the subsequent screen.