Better notification system
This implements an improvement to the current notification system.

The `UserMessage` differenciates three levels of severity:

* Info
* Warning
* Error

That helps to remove most of the `GenericError` messages which were
misused before as basic feedback.

Furthermore this adds a context, so we can later filter out if messages
are relevant to be visible or not.

Fixes: https://github.com/purebred-mua/purebred/issues/286