ci: add patch with custom build step
Add that patch that is used by haskell-ci (via --github-patches
option) to add our custom build steps.  As a result, we can just run
'haskell-ci regenerate' to adjust the GHC versions we test with,
without additional work to reinstate our build steps.

The amount of context in the patch was minimised to make more robust
against changes to haskell-ci output.  We insert our custom step
immediately before the "tests" step.