Use the widget name to avoid overflowing notifications
Some editors use real time validation. That is, per character we fire of
a thread to validate, but kill the thread if the next character is
received in less than 500ms.

However, when the editor widget gets input too fast (e.g. tmux driven
tests or to a lesser extend a copy & paste[1]) and the last character
finishes the input (e.g. `Enter` to switch to the next input) the
validation message "overflows" into the next widget.

Instead, use the context information on the user message, to only show
user messages for widgets still visible in the UI.

[1] There is specific copy & paste support (aka bracketed paste). Only
the quick succession of receiving input is meant in this context.

Fixes: https://github.com/purebred-mua/purebred/issues/412