plugins: add ConfigHook and new program entry point
Add ConfigHook for processing UserConfiguration at program
initialisation.  Add a new program entrypoint that is applied to a
[PluginDict] instead of a UserConfiguration, called
'purebredWithPlugins'.  I would like to make it the main entry
point, called 'purebred', but will initially leave it as a separate
symbol.

Also add the TweakConfig plugin that constructs a 'Plugin
ConfigHook' given a modification function.  Together with
'plugbredWithPlugins', this subsumes the 'purebred' entry point.