sendMail: return Bool to signal success/failure
To facilitate responding to send success or failure with different
actions, enhance the `Completable` type class with an associated
type family `CompletableResult`.  Update the
`'ComposeListOfattachments` instance to return a `Bool`.

This commit is just an interim step toward better handling of send
failure and support for "reject send" plugins.  It does not change
the behaviour of the program.