add support for conditional actions
Add `bindAction`, which makes it possible to decide how to act based
on the result from the previous action.  This function is similar to
`(>>=)` but `Action v ctx` does not have a lawful `Monad` instance.
(The description of the second action gets ignored, thus disagreeing
with the Applicative instance).

Also add `ifte` ("if-then-else"), an action combinator that chooses
the next action based on whether the previous action returned `True`
or `False`.  Because the arguments are constant Action values, we
can (and do) describe the behaviour of the combined action.