sendMail: improve error handling
The subroutines invoked by `sendMail` have inconsistent error
handling.  The result of `trySendAndCatch` gets written to the
AppState, but then processing continues.  Purebred always tries to
write and index a sent "file", even when sending failed.

Update `trySendAndCatch` to return a result in `MonadError Error m`.
Update `sendMail` to execute `trySendAndCatch` in the `runExceptT`.
As a result, if sending fails, `sendMail` returns *that* error and
does not attempt to write and index a "sent" file.