stay in compose view if send fails
If sending mail fails, we currently return to the thread index and
the mail is lost.  Update the send Action to be a conditional
action.  On success, return to the thread view.  But on failure,
don't switch view.  As a result, the user is able to take another
action.  They can try sending again, or save the mail as a draft.

Fixes: https://github.com/purebred-mua/purebred/issues/418
Related: https://github.com/purebred-mua/purebred/issues/294