uat: use separate config dir for each test
With test concurrency, some tests are interfering with each other.
In particular, the "send mail" action we use for testing writes the
send mail to a static filename in the config directory.

Update the test suite to create a separate config dir for every
test.  The config and precompiled binary are copied into each test's
config directory during setup to avoid recompilation.

I did consider hard-linking instead of copying.  This would avoid
copying the (quite large) purebred custom binary to each test's
config dir.  But I decided against this.  Future tests might modify
the configuration and we need to keep it "pristine" for each test.