drop support for GHC 8.6
Due to purebred-email-0.5 we now depend on a time >= 1.9, which
first shipped with GHC 8.8.  This breaks CI on GHC < 8.6 due to the
the build step that uses the "installed" constraint for bundled
libs.  So remove 8.6 from the CI matrix.

Because we're no longer testing 8.6, we should also drop official
support for it.  Bump base min bound to 4.13, the version that
shipped with GHC 8.8.1.