fix compilation with GHC 9.0
As of GHC 9.0, record field selectors preserve the user-written
order of quantified type variables.  This makes some of our record
selectors in compatible with how they are used elsewhere in the
program.

For the ResourceSpec and EntityCommand fields, it suffices to remove
the quantification.

For the plugin hook types, it is necessary to write each getter as a
standalone function with the correct type.