merge purebred-dump-keybinding into main program
The `purebred-dump-keybindings` program has to initialise the
configuration, which duplicates behaviour from the main program.
But there is an important difference: `purebred-dump-keybinding`
does not load user configuration or run config plugin hooks that
could customise the keybindings. So it can only show the default
keybindings.

Move this "dump keybindings" behaviour into the main program,
triggered by the `--dump-keybindings` flag, and remove the
`purebred-dump-keybinding` program.  As a result, there are fewer
programs to update when the API changes, and the
exe:purebred-dump-keybindings build target cannot be missed by
mistake.

Fixes: https://github.com/purebred-mua/purebred/issues/450