nix: stick with building purebred for x86 for now
The flake-utils allow us to build purebred for non-x86-64 systems.
However at the beginning we don't support these platforms. Part of the
downside is that we're also snapping up any potential broken package in
the nixos-unstable channel. Granted, we could pin to a specific
revision, but the channel is mostly fine for x86_64.

Drive-by add on to also run `nix flake check` in CI.