uat: avoid trailing slash in cp command
On FreeBSD (perhaps other OSes) copying a directory with trailing
slash causes cp(1) to copy the *contents* of that directory to the
target location.  This causes failure of the subsequent find(1)
command and, as a consequence, test setup fails for all tests.

Remove the trailing slash and use the POSIX '-R' option instead of
non-standard '-r'.