ui: simplify 'toggle'
Simplify the Toggleable class and in particular the 'toggle' effect.
The previous approach was "two phase" - first find the index of the
current item, then toggle it.  We now avoid this by way of the
'listSelectedElementL' Traversal, which targets the currently
selected item.

`listSelectedElementL` has been merged into Brick itself, but is not
yet released.  We can migrate to the Brick-supplied version after it
has been released.