ui: toggledItemsL: target "inner" values
The toggledItemsL targets the whole `Toggleable a`, which includes
the "is toggled" `Bool`.  This makes it an unlawful `Traversal` -
you can change the state that says which elements are targeted.  It
is also unneeded.  Functions are forced to handle a (Bool, a), but
always discard the first element.

Fix these problem by changing 'toggledItemsL' to target the inner
`a` values of the `Toggleable a` list items.  This makes it a lawful
`Traversal` and simplifies the traversal functions.