add uniform BodyPresentation type
Overview
--------

Add the `BodyPresentation` type.  Arbitrary data types could be
lifted into a `BodyPresentation`.  The `BodyPresentation` defines,
in a "source-specific" way, how to perform substring searches, how
to build the widget, and computes the scroll offsets on behalf of
the UI.

Put another way, the UI sees all `BodyPresentation` values as the
same, even though they may be very different "Under the hood".  For
the current time, we unconditionally construct a `MailBody` (as
before) and lift it into a `BodyPresentation` (new shiny).  Later we
can implement `BodyPresentation` for other "source" types, e.g. HTML
or Pandoc.  The configuration interfaces to handle conversion and
converter selection are not yet worked out, but this commit
establishes the presentation interface.

The core types of the machinery are defined in
`Purebred.Types.Presentation`.  The `MailBody`-specific
implementation is in `Purebred.Types.Presentation.MailBody`.
Existing `MailBody` types and functions were gathered here from
around the codebase.

Use of render cache
-------------------

The "scroll to match" behaviour is no longer implemented via scroll
requests generated by Actions.  The relevant actions now merely
update the match index in the program state.  Scrolling is
accomplished via Brick "visibility requests" generated while
composing the UI.

This requires use of the render cache.  Without the render cache,
Brick "sees" the visibility requests at all times, which "clamps"
the scrollable bounds to those containing the region requested to be
visible.  However, the render cache does not "remember" visibility
requests.  (This is by design; see Brick's Changelog.)  By using
`cached`, the "inner" widget's visibility request (if any) causes
the "scroll to match" on the initial render only.  The user is then
free to scroll anywhere in the viewport.

We invalidate the render cache on the following events:

- Search term applied, changed, or cleared
- Match index changed (i.e. cycling through matches)
- MailView entity changes

The upsides of this implementation detail are several:

- Simpler Action code.  We no longer need to peek into application
  state to emit scroll requests.

- The scroll offset is produced by the `BodyPresentation`; the
  behaviour is no longer coupled to a specific data type.

- Performance improvements by virtue of render cache.  Previously
  even a mere scroll up/down would cause the whole entity to be
  re-rendered.

Tests
-----

Several `MailBody` functions are no longer exported.  There were a
couple of basic tests for parsing and substring search.  I removed
these.  We will instead rely on the UAT suite.