Allow to expand receipients from addressbook
This implements an addressbook API and a way to expand addresses based
on addressbook entries. The default retriever is based on the Mutt
`alias` file using nick names.

The retrieval of Mutt aliases is not enabled by default, due to the fact
that we need to use an absolute path to the alias file. Guessworking the
possible location of the file goes against the current nature of being
very explicit in the configuration.

Fixes https://github.com/purebred-mua/purebred/issues/17